Move preview specific utility functions into separate file.
